restructure controller endpoints
